Merge develop_joel into master

+++ src/data_structures/APR/iterators/GenIterator.hpp
@@ -27,21 +27,21 @@ public:
 //    operator uint64_t() {return 0;};
 
     //helpers
-    inline unsigned int x_nearest_pixel( unsigned int level, int x);
-    inline float x_global( unsigned int level, int x);
-    inline unsigned int y_nearest_pixel( unsigned int level, int y);
-    inline float y_global( unsigned int level, int y);
-    inline unsigned int z_nearest_pixel( unsigned int level, int z);
-    inline float z_global( unsigned int level, int z);
+    inline int x_nearest_pixel(int level, int x);
+    inline float x_global(int level, int x);
+    inline int y_nearest_pixel(int level, int y);
+    inline float y_global(int level, int y);
+    inline int z_nearest_pixel(int level, int z);
+    inline float z_global(int level, int z);
 
-    unsigned int level_size( unsigned int level);
+    int level_size(int level);
 
-    unsigned int org_dims(int dim) const { return genInfo->org_dims[dim]; }
+    int org_dims(int dim) const { return genInfo->org_dims[dim]; }
 
-    unsigned int number_dimensions(){return genInfo->number_dimensions;};
+    int number_dimensions(){return genInfo->number_dimensions;};
 
-    uint16_t level_min();
-    uint16_t level_max();
+    int level_min();
+    int level_max();
 
     inline int x_num(const unsigned int level){
         return genInfo->x_num[level];
@@ -63,36 +63,36 @@ public:
 inline uint64_t GenIterator::total_number_particles() {return genInfo->total_number_particles;}
 
 
-inline unsigned int GenIterator::level_size(const unsigned int level){
+inline int GenIterator::level_size(const int level){
     return genInfo->level_size[level];
 }
 
-inline unsigned int GenIterator::x_nearest_pixel(const unsigned int level,const int x ){
-    return (unsigned int) floor((x+0.5)*level_size(level));
+inline int GenIterator::x_nearest_pixel(const int level,const int x ){
+    return (int) floor((x+0.5)*level_size(level));
 }
 
-inline float GenIterator::x_global(const unsigned int level,const int x) {
+inline float GenIterator::x_global(const int level,const int x) {
     return  (x+0.5f)*level_size(level);
 }
 
-inline unsigned int GenIterator::y_nearest_pixel(const unsigned int level,const int y ){
-    return (unsigned int) floor((y+0.5)*level_size(level));
+inline int GenIterator::y_nearest_pixel(const int level,const int y ){
+    return (int) floor((y+0.5)*level_size(level));
 }
 
-inline float GenIterator::y_global(const unsigned int level,const int y) {
+inline float GenIterator::y_global(const int level,const int y) {
     return  (y+0.5f)*level_size(level);
 }
 
-inline unsigned int GenIterator::z_nearest_pixel(const unsigned int level,const int z ){
-    return (unsigned int) floor((z+0.5)*level_size(level));
+inline int GenIterator::z_nearest_pixel(const int level,const int z ){
+    return (int) floor((z+0.5)*level_size(level));
 }
 
-inline float GenIterator::z_global(const unsigned int level,const int z) {
+inline float GenIterator::z_global(const int level,const int z) {
     return  (z+0.5f)*level_size(level);
 }
 
-inline uint16_t GenIterator::level_min() {return genInfo->l_min;}
-inline uint16_t GenIterator::level_max() {return genInfo->l_max;}
+inline int GenIterator::level_min() {return genInfo->l_min;}
+inline int GenIterator::level_max() {return genInfo->l_max;}
 
 
 #endif //LIBAPR_GENITERATOR_HPP
